Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudinary>=1.44.2",
    "django>=6.0.5",
    "django-cloudinary-storage>=0.3.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "google-auth>=2.52.0",
    "gunicorn>=26.0.0",
    "psycopg[binary]>=3.3.4",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
